fix(cli): extract offline impression rotation helper
