+++ packages/remnic-cli/src/index.ts
@@ -148,8 +148,7 @@ import {
   shouldPreferIncomingOfflineRuntimeFile,
   summarizeOfflineSyncChangeset,
   summarizeOfflineSyncPendingChanges,
-  summarizeOfflineSyncPendingFiles,
-  writeOfflineSyncState, drainPendingImpressionsForOfflineSync,
+  summarizeOfflineSyncPendingFiles, writeOfflineSyncState,
   type OfflineSyncApplyFileContentChunkResult,
   type OfflineSyncFileDigest,
   type OfflineSyncFileRecord,
@@ -188,11 +187,11 @@ import {
   readHeader,
   secureStoreDir,
 } from "@remnic/core/secure-store";
-import { LastRecallStore } from "@remnic/core/recall-state";
 // @remnic/export-weclone is an optional install surface (training:export
 // only uses it). Load lazily so the CLI works without it — see
 // optional-weclone-export.ts for the install-hint behaviour.
 import { loadWecloneExportModule } from "./optional-weclone-export.js";
+import { drainOfflineSyncImpressions, resolveOfflineImpressionRotation } from "./offline-impression-rotation.js";
 import type {
   BinaryLifecycleConfig,
 } from "@remnic/core";
@@ -9008,16 +9007,7 @@ export async function runOfflineSyncOnce(options: {
   userExcludeRegexps?: readonly RegExp[];
   /** Large files permanently skipped by the watch 3-strikes policy (#1786). */
   skipLargeFilePaths?: ReadonlySet<string>;
-  /**
-   * Recall-impression rotation bounds the daemon writer uses (#2033). A
-   * standalone CLI push drains pending impressions through its own
-   * `LastRecallStore`; passing the configured bounds keeps that drain's
-   * rotation identical to the writer's instead of silently reverting to
-   * `LastRecallStore` defaults.
-   */
-  impressionsRotateBytes: number;
-  impressionsRotateKeep: number;
-}): Promise<OfflineSyncRunResult> {
+} & { impressionsRotateBytes: number; impressionsRotateKeep: number }): Promise<OfflineSyncRunResult> {
   fs.mkdirSync(options.memoryDir, { recursive: true });
   let activeStatePath = options.statePath;
   let priorState = await readOfflineSyncState(activeStatePath);
@@ -9058,12 +9048,7 @@ export async function runOfflineSyncOnce(options: {
   const baseCapturedAt = priorState ? new Date(priorState.lastSyncedAt) : undefined;
   const storageIo = await createOfflineStorageIo(options.memoryDir);
   const localSourceId = localOfflineSourceId(options.memoryDir);
-  await drainPendingImpressionsForOfflineSync(() =>
-    new LastRecallStore(options.memoryDir, {
-      impressionsRotateBytes: options.impressionsRotateBytes,
-      impressionsRotateKeep: options.impressionsRotateKeep,
-    }).drainPendingImpressions(),
-  );
+  await drainOfflineSyncImpressions(options.memoryDir, options);
   const currentSnapshotForPush = await buildOfflineSyncSnapshotFromBase({
     root: options.memoryDir,
     sourceId: localSourceId,
@@ -9691,35 +9676,6 @@ function resolveOfflineSyncUserExcludes(rest: string[]): RegExp[] {
   return compileOfflineSyncExcludeGlobs(globs);
 }
 
-/**
- * Resolve the recall-impression rotation bounds the daemon writer uses so a
- * standalone CLI push drains through an identically-configured store (#2033).
- * Uses the same config parser as the writer; unparseable config surfaces loudly
- * rather than silently falling back to LastRecallStore defaults.
- */
-function resolveOfflineImpressionRotation(): {
-  impressionsRotateBytes: number;
-  impressionsRotateKeep: number;
-} {
-  const configPath = resolveConfigPath();
-  let raw: unknown;
-  try {
-    raw = fs.existsSync(configPath)
-      ? JSON.parse(fs.readFileSync(configPath, "utf8"))
-      : {};
-  } catch (error) {
-    if (error instanceof SyntaxError) {
-      throw new Error(`cannot read recall-impression rotation from ${configPath}: ${error.message}`);
-    }
-    throw error;
-  }
-  const config = parseConfig(resolveRemnicConfigRecord(raw));
-  return {
-    impressionsRotateBytes: config.recallImpressionsRotateBytes,
-    impressionsRotateKeep: config.recallImpressionsRotateKeep,
-  };
-}
-
 async function cmdOffline(action: string, rest: string[], json: boolean): Promise<void> {
   if (action === "help" || action === "--help" || action === "-h" || rest.includes("--help") || rest.includes("-h")) {
     console.log(`Usage: remnic offline <prepare|sync|status|watch> [options]
@@ -9748,7 +9704,7 @@ Environment fallbacks:
   const stateOverride = resolveRequiredValueFlag(rest, "--state");
   const statePathExplicit = stateOverride !== undefined;
   const userExcludeRegexps = resolveOfflineSyncUserExcludes(rest);
-  const impressionRotation = resolveOfflineImpressionRotation();
+  const impressionRotation = resolveOfflineImpressionRotation(resolveConfigPath());
   const needsRemote = action === "prepare" || action === "sync" || action === "watch";
   const remoteUrl = needsRemote
     ? resolveOfflineRemoteUrl(rest)
@@ -9835,8 +9791,7 @@ Environment fallbacks:
       statePath,
       statePathExplicit,
       userExcludeRegexps,
-      impressionsRotateBytes: impressionRotation.impressionsRotateBytes,
-      impressionsRotateKeep: impressionRotation.impressionsRotateKeep,
+      ...impressionRotation,
     });
     if (json) {
       console.log(JSON.stringify(offlineSyncResultJsonSummary(result), null, 2));
@@ -9920,8 +9875,7 @@ Environment fallbacks:
           statePath,
           statePathExplicit,
           userExcludeRegexps,
-          impressionsRotateBytes: impressionRotation.impressionsRotateBytes,
-          impressionsRotateKeep: impressionRotation.impressionsRotateKeep,
+          ...impressionRotation,
           skipLargeFilePaths: skippedLargeFiles,
         });
         const advanced = advanceOfflineLargeFileFailureCounts({
